+++ CHANGELOG.md
@@ -2,8 +2,13 @@
 
 ## v0.1.1
 
+[bugfix]
+- Delete url when the scope is not passed. Before, failed due to absent scope.
+
+[documentation]
 - remove use of pseudofolder from documentation as it is
 completely optional.
+
 
 ## v0.1.0